Support repeat-count for repeating operators
The variable `evil-repeat-count' is set to non-nil when `evil-repeat'
is called with a count. `evil-operator-range' uses the value of this
variable as count and afterwards resets `evil-repeat-count' to nil so
that count is used for the first operator only.

--HG--
extra : rebase_source : b12e0a59b2a17f6e1b66f281dba133de386b5b6b