Preserve current column in `evil-next-line' and `evil-previous-line'
This behavior is exactly what Emacs does -- but only if the last
command is one of `next-line' or `previous-line'. Furthermore Emacs
moves the cursor to the end/beginning of the last line reached if no
further movement is possible. In contrast ViM does never leave the
column at the buffer boundaries.