Add special handling of "0" command
"0" has two meanings in ViM. First it is the motion
'beginning-of-line, second it is a prefix-argument but only if it is
not the first number. The macro `evil-redirect-digit-argument' calls
either `digit-argument' or another command depending on whether "0" is
the first number typed.

Furthermore we have to bind all numbers to `digit-argument' in
evil-motion-state-map and remove those bindings from
`evil-suppress-map because otherwise the binding of the latter
function would overwrite the bindings to "0" in the motion map.