Redefine helper functions
Replace `evil-verify-around-point' with `evil-test-text', which splits
the EXPECTED argument in two: BEFORE for text preceding point and
AFTER for text succeeding point. These strings are not regular
expressions; use the optional BEFORE-PREDICATE and AFTER-PREDICATE
arguments to check for buffer boundaries.

Also replace `evil-test-editing' with `evil-test-macro' and
`evil-test-editing-clean' with `evil-test-buffer-edit'. Finally,
define `evil-test-code-buffer' in terms of `evil-test-buffer'
to avoid duplication.