Add `evil-define-key'
This assigns state bindings to a keymap. It is done by associating the
keymap with an auxiliary keymap, which is activated whenever the main
keymap is active. The association is stored in the list variable for
that state, e.g., `evil-normal-state-auxiliary-maps' for Normal state;
it might be possible to store it inside the main keymap instead, which
would work better with garbage collection.