Ensure that string key sequences are converted properly
`vconcat' is not sufficient to convert from string representation to
vector representation, as it does not handle modifier keys properly.
For example, (vconcat "\M-f") evaluates to [230], while (kbd "M-f")
returns a much larger value because the 2^27 bit in the character code
is used to indicate the meta key. The proper way is to call
`listify-key-sequence' first: (vconcat (listify-key-sequence "\M-f"))
returns the correct vector. Note that there is no safe way to convert
in the opposite direction, and the GNU Emacs Lisp Reference Manual
discourages the use of string representation.