Prevent buffer-switching in repeats
Constrain a repeat to a single buffer, aborting the repeat if the
buffer changes. To that end, add the variable `evil-repeat-buffer'
for keeping track of the current buffer. The motivation for this
is that most buffer changes are never intended to be repeated,
and when they are, they should consciously be recorded with "q".