Add PROFILER variable and "make profiler" rule
Let evil-tests.el instrument Evil for profiling. This is enabled with
"make profiler", which runs all tests in the terminal and then displays
profiling results, or with the PROFILER variable, which can be used
with any rule and specifies how the results are sorted (by call count,
average time or total time):

    make test PROFILER=call
    make test PROFILER=average
    make test PROFILER=total

Using the PROFILER variable with the "make emacs" rule creates a
results buffer with links to each function's source.