Record only if the *current* command is meant to be recorded
`evil-recording-current-command' is non-nil iff the *current* command
should be recorded. This is important in insert mode where
`evil-recording-repeat' is always non-nil but certain commands should
still be ignored. This is in particular necessary for the adviced
function `read-key-sequence' which is called when using input-methods,
but the corresponding commands should be ignored.