Fix selection of paragraph text objects
Ensure that the "vap" selects a paragraph in the same way as "dap".
When selecting multiple text objects in Visual state, the first object
is selected from the current position of point, and further objects
are selected by moving point outside the selection. This approach
replaces the backwards-looking heuristic from commit 71a4b8f.