Narrow to opening parenthesis in show-paren-function advice (fixes #43)
Emacs' show-paren-function can't highlight the second pair of parentheses in

(text1)(text2)

correctly if point is placed on the 2nd opening parenthesis because the first
pair is highlighted instead. In order to enforce highlighting of the second
pair in this situation we narrow the active region to the opening parenthesis.