Add `evil-visual-end-mark' function for special mark '>
Formerly the visual selection mark '> has been placed at the first
character after the selection because it was mapped directly to
`evil-visual-end'. Especially in visual line state this confused the
ex commands :'> and the interactive highlighting of :s. The former
moved point to the first line after the selection instead of the last
line of the selection. The latter also highlighted matches in the
first unselected line. This patch corrects the behavior by contracting
the visual region first before determining the position of '>.