Fix single lines as ranges
For an Ex command, ":123foo" is equivalent to ":123,123foo".
For an Emacs command, it is equivalent to "123 M-x foo".
For both, ".foo" is equivalent to ".,.foo", "$foo" to "$,$foo",
and so on; an Ex command receives the range via `evil-ex-range',
while an Emacs command perceives it as an active region.