Ensure `evil-shell-command' recognizes non-range case from ex correctly (fixes #77)
If the last line of a buffer is empty, i.e., the buffer ends with a
newline character, and point is placed on that line the command
:.!<cmd> calls `evil-shell-command' with an empty range. The old
behavior falsely interprets this case as :!<cmd>, i.e., as no range
given. The new behavior is to check explicitly if `evil-ex-range' is
nil.