Brackets in ex command bindings enclose optional command part
The old behavior was to enclose the obligatory part with brackets,
e.g., [e]dit defines e as an alias for edit. The new behavior is to
enclose the optional part, i.e., e[dit]. This form corresponds to the
style used in ViM. This is achieved by adapting `evil-ex-define-cmd'
only.