Quote functions with `function'
According to the GNU Emacs Lisp Reference Manual, section 12.7,
#' is a short-hand for `function' and serves as a note to the
compiler that the argument can safely be compiled. It also works
as a sort of comment when quoting the name of a function:

    #'SYMBOL = (function SYMBOL) = (quote SYMBOL) = 'SYMBOL

Since `lambda' is defined in terms of `function', the compiler will
compile unquoted `lambda' expressions anyway. Still, this convention
makes the source code more readable, and is similar to the usage of
#' in Common Lisp.