Make ex completion use `completion-at-point' framework.
That patch changes the definition of `evil-ex-define-argument-type' to
be usable with the completion-at-point framework. Thus, besides
activating and deactivating the argument handler, the completion of
arguments must be implemented according to the variable
`completion-at-point-functions'. The function `evil-ex-complete' is
removed in favor for `evil-ex-completion-at-point'.

The new framework allows for better selective programmed completion.