Fix handling of overlapping windows in `evil-ex-hl-update-highlights'
The old implementation incorrectly handled each match for each window
in which that match is visible. This resulted in multiple calls of the
match-hook for each match which in turn showed duplicate replacements
in interactive substitute. The new behavior is to collect all visible
ranges first and compute their union afterwards. The update of the
highlighting is then done w.r.t. to this union, hence each match that
is visible in at least one window is handled exactly once.
This fixes #104.