Ensure a" on the closing quote of a string selects the string
If point is at the closing quote of a string and this is indeed a
string in the sense of the current syntax table, then
`evil-paren-range' calls `evil-regexp-range' with those string
delimiters (because (point) is inside a string). But the opening and
the closing character are identical `evil-regexp-range' considers this
closing quote as an opening quote if there is another quote following
in that line. Thus not the string is selected but the text between the
closing quote and the following quote.

This patch checks whether the actual closing quote is indeed the
closing character of a string by testing if the following character is
still in a string and only if this test succeeds calls
`evil-regexp-range'. Note that this is only a quick fix and should not
considered a stable solution!

--HG--
extra : rebase_source : b85ae9b530d2d9938bcad8261b51383160283044