Add `evil-with-adjust-cursor' macro
Several motions wrap their code within this macro instead of calling
`evil-adjust-cursor'. This macro works by narrowing the buffer to (1-
(point-max)) if the buffer ends in a newline character. Thus the
motion does not reach the original (point-max) at all. This has the
advantage that the cursor has not to be moved back the correct column
in the previous line. This addresses issue #119.