Temporarily show invisible text when handling insert repeat count
The function `evil-cleanup-insert-state' uses `current-column' and
`move-to-column' to find the correct insertion position of the text to
be inserted. Those functions usually ignore invisible text. In the
case of text being invisible because of hideshow mode this text should
not be ignored and the insertion should take place at the invisible
lines, too. Thus we temporarily remove 'hs from
`buffer-invisibility-spec' which causes those lines to be visible and
being recognized by `move-to-column'. This addresses issue #128.