Remove spaces after opening an empty line
If a new line has been opened with o or O and nothing has been
inserted, this line consists only of whitespaces according to the
indentation. In this case the whitespaces are removed leaving a
completely empty line. This is achieved by a special post-command-hook
registered by `evil-open-below' and `evil-open-above'.
This fixes issue #132.