Ensure :wq does not ask for unmodified buffers
`evil-save-and-close' called `evil-write' with the full region
(point-min) to (point-max). The problem is that an explicit range
causes `evil-write' to always ask for the file to be overwritten
because an explicit region is never considered to be equivalent to the
full buffer content. Now `evil-write' is called with the region set to
nil which causes the full buffer content to be saved to its file only
if modified.