Add `evil-track-eol' variable
If this variable is non-nil (the default) a line move (motion j or k)
after an end-of-line (motion $) keeps point at the end of the line.
The Emacs customization variable `track-eol' is not sufficient because
of the different interpretation of end-of-line of plain Emacs and
Evil. This is achieved by setting `temporary-goal-column' to a large
number and predicting that `this-command' is `next-line' when
`evil-end-of-line' is called. This addresses issue #120.