Add `evil-apply-on-rectangle'
This is the same as `apply-on-rectangle' but may extend the region to
the end of each line. This happens if the last command has been a line
motion and temporary-goal-column is very large. The the actual end
column of the operation is that largest column of the covered block.
This addresses issue #120.