Fix `evil-ex-search-next' at eob or bob
Calling this command starts with a call to either `forward-char' or
`backward-char' depending on the direction of search to guarantee that
a new match, not the one at point, is found next. If point is at the
buffer boundaries this is not necessary because the search would wrap
anyway. This fixes issue #156.