Make `evil-intercept-maps' and `evil-overriding-maps' customizable
The former initialization used `evil-delay'. The problem with this
approach is that initialization cannot be reverted.

The new approach uses the variables `evil-pending-*-maps' to store
delayed overriding/intercept maps which are rechecked from a single
after-load-function. Furthermore the `override-state' and
`intercept-state' setting of keymaps that are removed from
`evil-overriding-maps' resp. `evil-intercept-maps' are removed (set to
nil), which effectively disables the override/intercept state.

The handling is done via a :set function of the corresponding
customization variables.

The initialization of those custom variables is delayed until evil is
loaded using `evil-after-load-hook'.