Update X selection according to visual state region
When *selecting* some visual region then the X selection should be
updated according to the values of `x-select-enable-primary' and
`x-select-enable-clipboard'.

The default Emacs behavior is to update the X selection whenever the
mark is changed if `select-active-regions' is non-nil. This usually
interferes with Evil visual state because then the visual selection
does not correspond to the Emacs region. This patch sets this variable
to temporarily to nil during visual state so Evil takes responsibility
for the X selection. The X selection is updated in
`evil-visual-post-command' by calling `x-select-text' with the
appropriate values.

This update does not work in block visual state.

This addresses issue #110.
This addresses issue #158.