Handle zero size matches in `evil-ex-substitute' on whole lines
If the match is of zero size and at the beginning of a line then the
first line after the select region matches two when using
`perform-replace'. The reason is that a zero size match in the case
does not exceed the region.

Therefore the region is reduced by one character before calling
`perform-replace' if the last character in the region is a newline.
This corresponds to the interactive highlighting.

This addresses issue #175.